Added utils unit tests
