+++ setup.cfg
@@ -15,6 +15,7 @@ packages = find:
 install_requires =
     importlib-metadata; python_version<"3.8"
     ruamel.yaml
+    jinja2
     PyYAML
 
 [options.package_data]
